Extract some shared code into `FileDataExtractor`
